AquaticPrime: Add `plist` method
Working version of a function that generates a license plist XML string
from input data.

Made a complete and utter mess of the code, and the two tests in this
commit aren't really tests, they're just a way of executing the
function.

The function takes a serializable data type, sends the data to the
`sign` function, and produces an XML plist string of the input data plus
a field for the signature.

In order to accomplish this, I ended up with a kind of roundabout data
manipulation.

1. We start with the user data type.
2. That then gets serialised to an XML plist using the 'plist' crate's
   'serde' serializer.
3. The serialized XML is then deserialized to a `Plist` instance.
4. The dict is extracted from the `Plist` and fed to `sign()` get a
   signature from the data.
5. We add the signature to the dict as a `Plist::Data`, in accordance
   with the Aquatic Prime license file format.
6. The full dict including the signature entry is serialized to XML and
   returned from the function.

I ended up with this circuitous manipulation because I wanted to be able
to accept an arbitrary struct as input data, and serializing it seemed
to be the easiest way to get a plist dict from it. But we can't just use
that serialized result, because it doesn't contain the signature.
Getting a `Plist` from the data seemed to be the right way to go, as the
signature entry needs to be a plist `data` type, so we can't just insert
it into a `<String, String>` `HashMap`.

Now that it works, I'll be cleaning all of this up in further commits.