Add migrations for `purchasers` table
A database table to hold purchaser information so we can re-generate
licenses in case purchasers lose their license keys.

Needed to use a trigger to update the `updated_at` field on `UPDATE` as
you can't do `ON UPDATE UTC_TIMESTAMP()` in the column definition

> You can not specify UTC_TIMESTAMP as default to specify automatic
> properties

(https://dba.stackexchange.com/questions/20217/mysql-set-utc-time-as-default-timestamp/21619#21619)

Weirdly, the trigger isn't working when applying the migration with

    $ migrant apply

but it is working when running

    $ mysql -u user dome_key < migrations/20181108234653_create-purchasers/up.sql

Not sure what's going on there, but 'migrant' appears to have trouble
realising there are errors coming back from MySQL and executes the
migrations regardless. It also doesn't print syntax error messages from
MySQL which is very inconvenient. Migrant seemed to be the most advanced
migration CLI on crates.io, and I was hoping to use a Rust program for
this, but for simplicity, I'm thinking I'll have to go with a different
migration runner. Considering https://github.com/golang-migrate/migrate.