.env.sample: Add `migrate` function
Wraps the `migrate` command (https://github.com/golang-migrate/migrate).
The command requires you to pass in a bunch of shell options that should
be constant for an application.

Encode those constants in this wrapper function. Now all you need to do
is run:

    $ migrate migration_name

and the correct migrations will be generated.

The migrations get created in the current directory. I tried passing

    -source file://migrations

,

    -source file://./migrations

, and

    -path migrations

but no combination seemed to put the generated migration files in the
migrations/ directory. Settled on moving the files manually in the
helper function.