20181109031633_create_purchasers.up.sql: Remove `DELIMITER` lines
Running

    $ mysql -u user dome_key < 20181109031633_create_purchasers.up.sql

worked just fine, but running the migration through 'migrate' produced
this error:

    (details: Error 1064: You have an error in your SQL syntax; check the manual that corresponds to your MariaDB server version for the right syntax to use near 'DELIMITER $$
    CREATE TRIGGER purchasers_updated_at
    BEFORE UPDATE
    ON purchasers FO' at line 1)

Probably the same problem that 'migrant' had, but it just didn't give me
an error message.

After having to fiddle and mess around with different parts of the
migration for a while, it finally turned out that the `DELIMITER` seemed
to be the problem.

Got rid of it as well as the `BEGIN`/`END` that depended on it. Looks
like our timestamp updater still works even without `BEGIN`, though, so
that's good. When I had first written the code, I figured I should use
`BEGIN` because that was how it was written in a couple examples I had
seen.