Purchaser: Add `generate_secret()` method
This new method generates a secret, which is a SHA1 digest of the
purchaser's name, email, and a random integer.

In order to use the `hexdigest()` method in the 'sha1' crate, I needed
to add the feature `std`
(https://docs.rs/sha1/0.6.0/sha1/struct.Sha1.html#method.hexdigest).

Needed to change the `secret` field to a `String` because otherwise the
generated digest string doesn't have a long enough lifetime to assign to
it.

Update `with_secret()` to use the new `String` type.

Update `insert()` to correctly handle the `Option` in `secret`.