Add 500 error page with generator
Base structure copied from `404.html`. Python script based on the
`generate_homebrew_formula.py` script in the main DomeKey repository.

The filename comes from the Apache server configuration.

Generate the 500 page because we can't rely on dependencies. This gets
the CSS and logo and includes them inline in the HTML page.

Thanks to this answer for explaining how to get a byte string from a
file to send to the base64 encoder:
https://stackoverflow.com/questions/45482272/typeerror-a-bytes-like-object-is-required-not-str-python-2-to-3/45482834#45482834