index.html: Increase minimum OS requirement
Just tested the program on 10.8 and it didn't work. It ran just fine,
but didn't run map actions, and I'm not sure why. Used the release
build, so I didn't debug it at all.

Since I'm planning an imminent public release, just set the minimum to
an OS version that we know for sure works.

In the future, I'd still like to get the program working on 10.8. We can
always bring down the minimum OS version.