index.html: Add demo video
Not committing the video files. They're only 3 & 4 megabytes each, but
that still seems a bit too much to keep in version control.

Add the video to the home page and center it in the content area. We
give it a negative margin to center it since it's wider than the
container.