Fix video width on small screen widths
When the screen width gets small enough to clip the video, reduce the
width of the video to the content width.

Couldn't use the same larger-than-content-width layout, at least not
easily, because the percentage would need to be dependent on the window
width, not the width of the video element's container.