Rename `bin/license-generator.rs` to bin/fulfillment.rs`
Renames the binary. The name "license-generator" no longer seems apt at
this point. That binary doesn't actually generate a license, it serves
as the webhook endpoint for Paddle's license fulfillment. Our `license`
binary handles the actual license generation.

The URL I'll be using for the webhook is `/fulfillment`, so this also
makes sense from that perspective.