Add deployment script
To deploy, run

    $ make deploy

This checks prerequisites and runs `scripts/deploy.sh`.

Makefile:
Include a new target that builds the Rust binaries for Linux using
Docker.

The script `rsync`s the website files to the server, moves some files
around on the server, and runs the database migrations.

Currently I'm getting an error at the database migration stage, as it
seems I can't use `UTC_TIMESTAMP()` as a default column value. Need to
correct that to get a clean deployment.