license-generator: Use compile-time environment variables
For whatever reason, the environment variables in my `.htaccess` weren't
getting passed to my FastCGI executables on my production server.

Wasn't sure how to get them passed to the programs, so decided instead
to compile the env variables into the binaries.

To do that, we source the environment file before building the release
builds in the Docker container.