Prevent selection of '$' in shell commands
Now that it takes two commands to install the program, make it easier
for visitors to copy-paste the commands into their shell by preventing
selection of the '$' prompt character.

This allows us to keep the prompt to indicate that these are shell
commands and still make copy-pasting easier.

As a side note, decided not to add the `brew services` command to the
instructions on the thank-you page because if DomeKey is already
running when you license it, you need to restart it for the license to
take effect.